Experiment with stub based typing
Mypy's stubgen tool can generate type stups (.pyi files) for
oslo_versionedobject. Every o.vo field class can potentially has an
AUTO_TYPE member where the type of the field is specified in o.vo. The
stubgen generates this as

AUTO_TYPE: Any = ...

We could manually edit the stub and define the type of the AUTO_TYPE to
a python type (instead of the o.vo type runtime).

While this seemed to help not to hardcode the type mapping between o.vo
field types and python types it also has drawbacks:
* stubs need to be re-generated to pick up new o.vo field types in the
  future
* manually edited stubs does not play well with auto generation of the
  stubs. We might need to store diffs and apply them on the regenerated
  stubs.
* in general the mapping between o.vo field types and python type are
  hard due to o.vo value coercing mechanism. E.g. o.vo allows assigning
  string value to an IntegerField if the value can be converted to int
  via int(). Code bases that does not rely on this coercion would like
  to use the int as a type for an IntegerField to catch errors. While
  codebases that rely on the coercion needs to define the field type as
  str. So we need to make type mapping configurable. But there are a lot
  easier way for that than editing a autogenerated stub.

This seems to be a dead-end right now.