Make type inference based on type stubs
A partial oslo.versionedobjects stub is added to have a place where the
o.vo field types can be defined. A new type checking only member MYPY_TYPE
is added to the o.vo field types to indicate to mypy what type the field
should get inferred.

This makes it possible to define types and type aliases with a full typing
syntax in the stub file.